feat(cli)!: rename `bench compat` -> `bench hub` (env-hub compatibility)
Per product direction: "compat" is the wrong word when the goal is being
compatible with every environment hub. Renamed the whole surface to a
hub-centric noun:

- CLI: `bench compat harbor-registry` -> `bench hub check` (group + a
  shorter subcommand). cli/compat.py -> cli/hub.py, register_compat ->
  register_hub.
- Package: src/benchflow/compat/ -> src/benchflow/hub/ (the Harbor-check
  package; harbor_registry.py keeps its harbor-specific fn names — no
  speculative abstraction). Cache dir .cache/compat/harbor -> .cache/hub/harbor.
- Updated the importers (cli, integration evidence checker), the
  cli-docs-drift guard, and docs/reference/cli.md.

IMPORTANT: left the task.md frontmatter `benchflow.compat.*` namespace
(provenance for compat-imported tasks, in task/_document_parse/export/imports)
UNTOUCHED — that is a separate task-standard concept, not this command.

Clean rename (no alias): `bench compat` is a v0.6-only command with no
public users yet. Full suite 4041 passed; ruff + ty clean.
